# src/main/java/pro/sky/finalprojectsky/dto/UserDto.java
package pro.sky.finalprojectsky.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

import javax.validation.constraints.Email;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Pattern;
import javax.validation.constraints.Size;

import static pro.sky.finalprojectsky.constant.Regexp.EMAIL_REGEXP;
import static pro.sky.finalprojectsky.constant.Regexp.PHONE_REGEXP;

@Data
public class UserDto {


    private int id;

    @Email(regexp = EMAIL_REGEXP)
    @Schema(example = "user@user.ru")
    private String email;

    @NotBlank
    @Size(min = 3)
    private String firstName;

    @NotBlank
    @Size(min = 3)
    private String lastName;

    @Pattern(regexp = PHONE_REGEXP)
    private String phone;
}